requirements.txt
Flask
pandas
scikit-learn
flask-socketio
simple-websocket
google-adk
gunicorn
python-dotenv
google-cloud-aiplatform
